+++ lib/Loc_tree.mli
@@ -13,5 +13,5 @@ open Migrate_ast
 
 include Non_overlapping_interval_tree.S with type itv = Location.t
 
-val of_ast : 'a Mapper.fragment -> 'a -> Source.t -> t * Location.t list
+val of_ast : 'a Traverse.fragment -> 'a -> Source.t -> t * Location.t list
 (** Use Ast_mapper to collect all locs in ast, and create a tree of them. *)